separate actions from slices
