+++ mobile/src/redux/reducers/articles-screen/adapters/articles.ts
