+++ mobile/src/redux/reducers/articles-screen/selectors/index.ts
@@ -1,4 +1,4 @@
-import { RootState } from "../..";
+import { RootState } from "../../..";
 import { articlesAdapter } from "../adapters/articles";
 
 export const { selectAll: selectArticles } = articlesAdapter.getSelectors(
