+++ mobile/src/redux/reducers/articles-screen/slice.ts
@@ -1,8 +1,8 @@
-import { createAsyncThunk, createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { useSelector } from "react-redux";
 
-import { fetchV2 } from "../../utils/fetch";
-import { RootState } from "..";
+import { RootState } from "../..";
+import { fetchArticle, fetchArticles } from "../../actions/articles-screen";
 import { articlesAdapter } from "./adapters/articles";
 import { selectArticles } from "./selectors";
 
@@ -15,32 +15,6 @@ const initialState = articlesAdapter.getInitialState<ArticlesScreenState>({
   status: "idle",
   error: null,
 });
-
-export const fetchArticles = createAsyncThunk("articlesScreen/fetchArticles", async () => {
-  try {
-    const articles = await fetchV2("data:articles/list.c.json", {
-      query: [["language", "en"]],
-    });
-    return articles;
-  } catch (error: any) {
-    return error.message;
-  }
-});
-
-export const fetchArticle = createAsyncThunk(
-  "articlesScreen/fetchArticle",
-  async (slug: string) => {
-    try {
-      const article = await fetchV2(`data:articles/:slug.json`, {
-        params: { slug },
-        query: [["language", "en"]],
-      });
-      return article;
-    } catch (error: any) {
-      return error.message;
-    }
-  },
-);
 
 const articlesScreenSlice = createSlice({
   name: "articlesScreen",
