+++ mobile/src/redux/reducers/contribute-screen/adapters/contributions.ts
