+++ mobile/src/redux/reducers/contribute-screen/adapters/filters.ts
