+++ mobile/src/redux/reducers/general/slice.ts
@@ -3,7 +3,7 @@ import Debounce from "debounce";
 import { Appearance } from "react-native";
 import { useSelector } from "react-redux";
 
-import { AppDispatch, RootState } from "..";
+import { AppDispatch, RootState } from "../..";
 
 interface GeneralState {
   theme: "dark" | "light";
@@ -24,23 +24,6 @@ export const getThemeFromSystem = (
 const initialState: GeneralState = {
   theme: getThemeFromSystem(),
 };
-
-export const init = createAsyncThunk<
-  any,
-  any,
-  {
-    dispatch: AppDispatch;
-    state: RootState;
-  }
->("general/init", (_, { getState, dispatch }) => {
-  const applyThemeDebounced: Appearance.AppearanceListener = Debounce(() => {
-    const theme = getThemeFromSystem();
-    const currentTheme = getState().general.theme;
-    if (theme === currentTheme) return;
-    dispatch(setTheme(theme));
-  }, 200);
-  Appearance.addChangeListener(applyThemeDebounced);
-});
 
 const generalSlice = createSlice({
   name: "general",
