+++ mobile/src/redux/reducers/learn-screen/adapters/documents.ts
