+++ mobile/src/redux/reducers/projects-screen/adapters/projects.ts
