+++ mobile/src/redux/reducers/projects-screen/selectors/index.ts
@@ -1,4 +1,4 @@
-import { RootState } from "../..";
+import { RootState } from "../../..";
 import { projectsAdapter } from "../adapters/projects";
 
 export const { selectAll: selectProjects } = projectsAdapter.getSelectors(
