+++ mobile/src/screens/articles/article-details/index.tsx
@@ -12,8 +12,9 @@ import { ErrorBoundary } from "../../../components/error-boundary";
 import { DZCodeLoading } from "../../../components/loading";
 import { TryAgain } from "../../../components/try-again";
 import { AppDispatch } from "../../../redux";
-import { fetchArticle, useArticlesSliceSelector } from "../../../redux/articles-screen/slice";
-import { useGeneralSliceSelector } from "../../../redux/general/slice";
+import { fetchArticle } from "../../../redux/actions/articles-screen";
+import { useArticlesSliceSelector } from "../../../redux/reducers/articles-screen/slice";
+import { useGeneralSliceSelector } from "../../../redux/reducers/general/slice";
 import { globalStyles } from "../../../styles/global";
 import { openLink } from "../../../utils/link";
 import { articleDetailsStyles } from "./styles";
