+++ mobile/src/screens/contribute/index.tsx
@@ -8,11 +8,11 @@ import { ErrorBoundary } from "../../components/error-boundary";
 import { DZCodeLoading } from "../../components/loading";
 import { TryAgain } from "../../components/try-again";
 import { AppDispatch } from "../../redux";
+import { fetchContributions } from "../../redux/actions/contribute-screen";
 import {
-  fetchContributions,
   updateFilterValue,
   useContributeSliceSelector,
-} from "../../redux/contribute-screen/slice";
+} from "../../redux/reducers/contribute-screen/slice";
 import { globalStyles } from "../../styles/global";
 import { CardItemMemoed } from "./card-item";
 import { contributeStyles } from "./styles";
