+++ mobile/src/screens/faq/index.tsx
@@ -5,8 +5,8 @@ import Markdown from "react-native-markdown-display";
 import { List, Text } from "react-native-paper";
 
 import { ErrorBoundary } from "../../components/error-boundary";
-import { useFaqSliceSelector } from "../../redux/faq-screen/slice";
-import { useGeneralSliceSelector } from "../../redux/general/slice";
+import { useFaqSliceSelector } from "../../redux/reducers/faq-screen/slice";
+import { useGeneralSliceSelector } from "../../redux/reducers/general/slice";
 import { globalStyles } from "../../styles/global";
 import { openLink } from "../../utils/link";
 import { faqStyles } from "./styles";
