+++ mobile/src/screens/learn/document-details/index.tsx
@@ -11,8 +11,9 @@ import { ErrorBoundary } from "../../../components/error-boundary";
 import { DZCodeLoading } from "../../../components/loading";
 import { TryAgain } from "../../../components/try-again";
 import { AppDispatch } from "../../../redux";
-import { useGeneralSliceSelector } from "../../../redux/general/slice";
-import { fetchDocument, useLearnSliceSelector } from "../../../redux/learn-screen/slice";
+import { fetchDocument } from "../../../redux/actions/learn-screen";
+import { useGeneralSliceSelector } from "../../../redux/reducers/general/slice";
+import { useLearnSliceSelector } from "../../../redux/reducers/learn-screen/slice";
 import { globalStyles } from "../../../styles/global";
 import { documentDetailsStyles } from "./styles";
